blade/api/requirements.txt
numpy==1.18.1
flask==1.1.2
Flask-Bootstrap==3.3.7.1
flask_appconfig==0.11.1
pymongo==3.10.1
flask-nav==0.6
Flask-WTF==0.14.3
pyyaml==5.3.1
Flask-Cors==3.0.9